Rewrite application in Flask (Python)
Dropbox dropped support for their v1 API and do not have an official SDK for their v2 API. I've also really been enjoying working in Python vs PHP, so it was a good excuse to migrate.

Updates to the README will be forthcoming, as well as removal of the PHP application.